requirements.txt
Flask>=3.0.0
pycryptodome>=3.19.0
numpy>=1.26.0

